# workspace/imports/inspiration-research/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Base API",
    "description": "Local-first personal knowledge base API",
    "version": "0.1.4"
  },
  "paths": {
...
    "/b-line/knowledge/extract": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Knowledge Extract",
        "description": "\u4ece\u6587\u672c\u4e2d\u63d0\u53d6\u7ed3\u6784\u5316\u673a\u5668\u77e5\u8bc6\u3002",
        "operationId": "knowledge_extract_b_line_knowledge_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/knowledge/search": {
      "get": {
        "tags": [
          "b-line"
        ],
        "summary": "Knowledge Search",
        "description": "\u641c\u7d22\u673a\u5668\u77e5\u8bc6\u6761\u76ee\u3002",
        "operationId": "knowledge_search_b_line_knowledge_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "unit_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Unit Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/knowledge/{knowledge_id}/evidence": {
      "get": {
        "tags": [
          "b-line"
        ],
        "summary": "Knowledge Evidence",
        "description": "\u4e3a\u6307\u5b9a\u77e5\u8bc6\u6761\u76ee\u751f\u6210\u53ef\u8ffd\u6eaf\u8bc1\u636e\u5305\u3002",
        "operationId": "knowledge_evidence_b_line_knowledge__knowledge_id__evidence_get",
        "parameters": [
          {
            "name": "knowledge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Knowledge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/agent/execute": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Agent Execute",
        "description": "\u901a\u8fc7 swarm-dev \u6280\u80fd\u6267\u884cAgent\u4efb\u52a1\u62c6\u89e3\u3002",
        "operationId": "agent_execute_b_line_agent_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/agent/decompose": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Agent Decompose",
        "description": "\u5c06\u9ad8\u7ea7\u76ee\u6807\u62c6\u89e3\u4e3a\u53ef\u6267\u884c\u7684\u5b50\u4efb\u52a1\u5e8f\u5217\uff08\u901a\u8fc7 swarm-dev task_decomposer\uff09\u3002",
        "operationId": "agent_decompose_b_line_agent_decompose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/agent/validate": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Agent Validate",
        "description": "\u9a8c\u8bc1\u4efb\u52a1\u6267\u884c\u7ed3\u679c\uff08\u901a\u8fc7 swarm-dev result_merger\uff09\u3002",
        "operationId": "agent_validate_b_line_agent_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/b-line/translate/assess": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Translate Assess",
        "description": "\u8bc4\u4f30\u7528\u6237\u5728\u77e5\u8bc6\u9886\u57df\u7684\u638c\u63e1\u5ea6\uff0c\u5224\u65ad\u662f\u5426\u8fbe\u5230\u673a\u5668\u8f6c\u8bd1\u9608\u503c\u3002",
        "operationId": "translate_assess_b_line_translate_assess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/translate": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Translate",
        "description": "\u5c06\u4eba\u7c7b\u77e5\u8bc6\u8f6c\u8bd1\u4e3a\u673a\u5668\u77e5\u8bc6\u6761\u76ee\u3002",
        "operationId": "translate_b_line_translate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/b-line/translate/publish": {
      "post": {
        "tags": [
          "b-line"
        ],
        "summary": "Translate Publish",
        "description": "\u53d1\u5e03\u673a\u5668\u77e5\u8bc6\u6761\u76ee\u5230\u77e5\u8bc6\u5e93\u4f9b\u4e0b\u6e38\u6d88\u8d39\u3002",
        "operationId": "translate_publish_b_line_translate_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}